+++ src/langchain_mcp_tools/__init__.py
@@ -3,4 +3,7 @@ from .langchain_mcp_tools import (
   convert_mcp_to_langchain_tools,
   McpServerCleanupFn,
   McpServersConfig,
+  McpServerCommandBasedConfig,
+  McpServerUrlBasedConfig,
+  SingleMcpServerConfig,
 )