+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+- Make SingleMcpServerConfig public (it used to be McpServerConfig)
+- Improve documentations
+
 
 ## [0.2.3] - 2025-04-22